requirements.txt
attrs==25.1.0
blinker==1.9.0
click==8.1.8
flasgger==0.9.7.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.1.1
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
